+++ packages/camera/camera_android/android/src/main/java/io/flutter/plugins/camera/CameraUtils.java
@@ -12,6 +12,7 @@ import android.hardware.camera2.CameraManager;
 import android.hardware.camera2.CameraMetadata;
 import io.flutter.embedding.engine.systemchannels.PlatformChannel;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -94,7 +95,27 @@ public final class CameraUtils {
   public static List<Map<String, Object>> getAvailableCameras(Activity activity)
       throws CameraAccessException {
     CameraManager cameraManager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
-    String[] cameraNames = cameraManager.getCameraIdList();
+    List<String> cameraNames = new ArrayList<>(Arrays.asList(cameraManager.getCameraIdList()));
+    List<Map<String, Object>> cameras = new ArrayList<>();
+
+    boolean expectingCamera = true;
+    int i = 0;
+
+    while (expectingCamera) {
+      try {
+        String cameraName = String.valueOf(i);
+        cameraNames.remove(cameraName);
+        CameraCharacteristics characteristics = cameraManager.getCameraCharacteristics(cameraName);
+        Map<String, Object> details = serializeCameraCharacteristics(cameraName, characteristics);
+        cameras.add(details);
+        i++;
+      } catch (Exception e) {
+        // retrieving Camera failed, most probably there is no other physical non-removable camera.
+        expectingCamera = false;
+      }
+    }
+
+
     List<Map<String, Object>> cameras = new ArrayList<>();
     for (String cameraName : cameraNames) {
       int cameraId;
@@ -107,26 +128,31 @@ public final class CameraUtils {
         continue;
       }
 
-      HashMap<String, Object> details = new HashMap<>();
       CameraCharacteristics characteristics = cameraManager.getCameraCharacteristics(cameraName);
-      details.put("name", cameraName);
-      int sensorOrientation = characteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
-      details.put("sensorOrientation", sensorOrientation);
-
-      int lensFacing = characteristics.get(CameraCharacteristics.LENS_FACING);
-      switch (lensFacing) {
-        case CameraMetadata.LENS_FACING_FRONT:
-          details.put("lensFacing", "front");
-          break;
-        case CameraMetadata.LENS_FACING_BACK:
-          details.put("lensFacing", "back");
-          break;
-        case CameraMetadata.LENS_FACING_EXTERNAL:
-          details.put("lensFacing", "external");
-          break;
-      }
+      Map<String, Object> details = serializeCameraCharacteristics(cameraName, characteristics);
       cameras.add(details);
     }
     return cameras;
+  }
+
+  private static Map<String, Object> serializeCameraCharacteristics(
+      String name, CameraCharacteristics cameraCharacteristics) {
+    HashMap<String, Object> details = new HashMap<>();
+    details.put("name", name);
+    int sensorOrientation = cameraCharacteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
+    details.put("sensorOrientation", sensorOrientation);
+    int lensFacing = cameraCharacteristics.get(CameraCharacteristics.LENS_FACING);
+    switch (lensFacing) {
+      case CameraMetadata.LENS_FACING_FRONT:
+        details.put("lensFacing", "front");
+        break;
+      case CameraMetadata.LENS_FACING_BACK:
+        details.put("lensFacing", "back");
+        break;
+      case CameraMetadata.LENS_FACING_EXTERNAL:
+        details.put("lensFacing", "external");
+        break;
+    }
+    return details;
   }
 }